Fix problem with very large prime bounds.
Large prime bounds larger than an unsigned long are now working more or
less correctly. This assumes that all the primes found by ECM fit in an
unsigned long: the very large primes just come from a cofactor that
happens to be a prime. On 32-bit machines, this will reach the limit.