fixed undefined behaviour with unbalanced splits.
In passing, this commit also introduces exact transfer counts for
scatter_mat and gather_mat, which are all based on Send/Recv pairs
anyway. In principle, we should be able to do them the Scatter/Gather
instead, but that seems both far away given the code organization, and
close to useless in this case.