implement tasks.sieve.qmax
logic: the old condition "stop if we have the requested number of
relations" remains. And that holds also when that number is
automatically guessed from the lpb.

On top of that, we add the constraint that no sieving WU is ever
submitted for q0 above qmax. Hence, in order to make sure that a q-range
is 100% sieved, the strategy entails setting rels_wanted to something
very large.