fix attempt for #21688
This gets rid of the timers that were computed with
getrusage(RUSAGE_SELF) by lack of something more precise. Now threads in
the thread pools do the work of collecting their own timing, and we
rebuild the tallied timings from that data.

Note that the thread wait times in the thread pool soars when we use
hyperthreading. I think that it is somehow a good sign (the other
hyperthread is busy).

It is plausible that the stats parsing in the python script will be
disturbed by the mildly changed stats line. TODO, then.