modified fake_rels ; should no longer need -v nor -sync
this version is still temporary, and composite special-qs probably don't
work.

Also, the algorithm is inadvertently changed for a different one, and
I'm willing to revert that: in this version, model relations are picked
exactly from the same special-q that was used to determine the number of
relations. That isn't really the spirit. (but on the other hand, clearly
the algorithm doesn't work too well then if we choose very different
special-qs for a given run).