ClientServerTask ; die if localhost clients die.
This effectively aborts the computation if a localhost client dies.
Overall, this idea of clients spawned by the server is deprecated in my
opinion, so it's not a big deal. Most often, when such a client dies,
it's something that gets in our way within tests (e.g. a test that never
returns because it waits for the timeout to expire, which is due to
happen if the real even is that the client has just died).

I think it's better that way (behaviour restricted to localhost client,
remember).  But it *might* be disruptive in certain cases.