ah, we need gzip, and it's not by default on opensuse...
somehow, this reveals the fact that we're unconditionally assuming that
the tools exist. Maybe we should adjust code such as
fopen_maybe_compressed to compile only the surface encodings that can
work on the current system. Not that much of a problem for gzip since
it's normally there, but the error message that we get when a .xz
filename is passed is definitely not pretty.