get rid of lrand48 completely, + rstate in rootfinding.
This patch exposes the random state object in the rootfinding api.

There's the question of the point to which it makes sense for this
object to re-surface. The rationale that is used here is that whenever
code does stuff for several p's, the random state is created on purpose
rather that needlessly exposed (typical example: get_alpha)

lrand48 is MT-Unsafe, has been an offender of many code scanners for
many years. It's time to go. (incidentally, this fixes CID 1453777 since
coverity is the last such tool to complain about lrand48 -- but the
issue is far older than that)