fixed bug reported by Robert Balfour:
https://sympa.inria.fr/sympa/arc/cado-nfs/2021-08/msg00011.html

The issue was that with a large value of nq (here 46656) it can
happen that the number of available "small q's" and their roots
is so low that the routine "next_comb" loops over the binomial(lq,k)
combinations. The fix was simply to detect the last combination
(which was planned in the code, but the return value of next_comb
was not used).