Compute factor base slices in a better way.
Prior to sieving, a slicing of the factor base is performed. Depending
on the logbase being used, the sets of primes with constant logp and
constant numbers of roots are split into roughly equal size and equal
weight pieces, in order for multithreading to go well.

Alas, the way it was done was somewhat stupid, and was actually hiding a
quadratic complexity algorithm: compute all slices from the start (each
endpoint requires a dichotomy on the cumulative weight distribution
function), and try with one more slice if any overflows.

Since the slices stretch to larger and larger sizes compared to the average
as the primes grow (and 1/p drops), this had the effect of hitting the
wall repeatedly until the "good" number of slices, each time redoing the
computation from the start.

This patch does two things:
 - compute these slice endpoints starting from the _last_ slice, not the
   first. This is because we expect failure to happen there, so let's
   detect it early.
 - based on how much a slice overflows, try to guess what a good "next
   number of slices" could be. This might still be off by a tiny bit,
   but seems to be generally right.