A tentative fix for #30023
When we're in fill_in_buckets_lowlevel AND first_reg is false
AND is_projective_like is true, then we have a projective prime in the
bucket-sieved range. With the fix of #30012, we want to deal with it, and
the code is correct, except that it checks for first_reg twice.

Without FIX_30012, the current code mistakenly avoids the branch that
has the "continue" statement if first_reg is false, which is clearly
wrong.