small technical fix (demise of the .badideals file)
we no longer need a .badideals (nor .badidealinfo) file, so let them be
gone for good.

What's slightly annoying, though, is that we do need to compute the unit
rank, and have it known to the python script. The most robust way that
I've found is unfortunately to keep the call to the badideals binary
(now called numbertheory_tool), and only retain that part of its
behaviour: provide the computation of the unit rank.