add new repo policy: sort file lists
ok, people might end up hating me. But I've had too many merge conflicts
caused by the usual policy of appending to the tail of the list, really.
In most cases, it was a better idea to put the new file at its right
place, in alphabetical order.

And we have a command-line tool to do the right thing anyway:
    ./scripts/check_file_lists.pl -w
(of course, maintaining the sort order with your favorite text editor
also works)