Delete old code
As the previous commit has shown, at some point in the past I decided to
ditch the old polyselect_hash mechanism altogether, and use the
polyselect_shash stuff instead.

The problem with #30040 was that by doing so, the replacement that I
came up with wasn't doing things correctly, and we ended up detecting
nonexisting collisions.

Since the old polyselect_hash code is now unused, let's delete it.