Some fun with ramified ideals
This tool is certainly not perfect, and shows that some more data should
probably be stored somewhere. Also, we might want to do the checks
internally with our code and not rely on an external tool for he checks
(although having both is good, of course!).

At this point, indexed relations on a c60 are easy to check.