Fix protecting macros for _mm256_cmpneq_epu32_mask
This intrinsic is avx512vl. For some reason it was in a branch that was
protected by avx512f && avx512dq. I don't know if there is a platform
that has f, dq, and not vl, but what I do know for sure is that because
of #30027, we forcibly disable vl on outdated compilers.