Fix wrong override of relation_ab::active_sides
This was all introduced by e25bbd5d1. Clearly, having active_sides
defined both in the base and derived class is wrong. It does make sense
in the base class, let's put it there.

The default setting of active_sides[0,1] to 0 and 1 will have to be
changed at some point.