got it -- sometimes transient bugs are worth looking into.
random matrix generation was guided by the dimensions of the
'significant part' of the matrix that it had to build (despite the
padded_nrows / padded_ncols argument names), and in some cases it meant
that it was missing some padding columns.

This whole layer of random matrix generation should be rewritten more
cleanly.