alpine: add gzip
the latest alpine release does away with the gzip package and uses the
embedded gzip that is found in busybox. This is well and good, but
unfortunately lcov sees it differently: it checks whether gzip -h says
something on stdout (which is what GNU gzip does). It does not like
seeing output on stderr only (which is what busybox gzip does).