Fix catastrophic bug caused by non-type safe C style printf
The GMP specific format strings have the effect of disabling the
compiler checks of the format string. Anything whose type is not
consistent with what the format string says (e.g. printing an uint64_t
with a %d) will cause the parsing of the varargs to go astray.

This bug has been here for long, but the only thing specific about the
current branch is that it (at last) enables dlog for platforms other
than amd64, which reveal the failure.

Of course we should be using fmt instead.