improve the logic to find the nearest parameter size
Previously we search for the exact digit size, then for the nearest multiple
of 5 (for size < 200 digits). Thus if we factored a c142 and we had a parameter
file for 143 digits, we still used the parameter file for 140 digits.

Now we search for the exact digit size, then for the nearest value in
[size_of_n - 3, size_of_n + 3].