Refactor duplicate implementations of {mpz_,}poly{modF,_scaled}
Bugs aside, we are also gaining in functionality since the multipleer
for parallel code now also goes to this mpz_polymodF type.

The C++ equivalent type is also added, following usual pattern.

Note that the implementation of mpz_polymodF is in mpz_poly.cpp ; there
is no mpz_polymodF.cpp. It's on purpose, because both are subscribed
into the mpz_poly_parallel multiplexed type.