Don't do import automatically in DoesImport.run() - call do_import() instead
Trying to do it automatically runs into problems with the order in which
instance variables get initialized in Task.run() and DoesImport.run().
With the current method resolution order, Task.run() calls
DoesImport.run() before self.merged_args is set up, and DoesImport.run()
requires those args to know the search path for the skewness binary.
It's probably better not to do to many things automatically - explicit
is better than implicit, or something.