change the concurrency logic in parallel tests
This was long overdue. It's still not entirely satisfactory, but we have
to fight dragons, here. Most importantly, the discovery on the hwloc
topology on a busy 512-core machine can easily take several seconds, and
this kills our timeouts. Short of a better solution, we're faking a
topology in order to alleviate this problem.

This ends up yielding a 10-fold speedup of the test suite run on a
512-core machine.