remove useless test
Interestingly, the _headers_ of the operator<<(mpz_srcptr) and friends
are in gmp.h anyway, but the implementation is classically in libgmpxx,
and we don't want to test if the distribution includes this shared
library or not.

So, in a way our gmpxx.hpp is useless, and we don't really need it. In
fact we never count on it.