make pi.size consistent across the low-level lingen computations.
Most of the matpoly member functions are likely to rely on pi.size being
correct (and tbh, they should also strive to _keep_ it correct). realloc
is one of them, for example, and it is used in the inner loop.

No bug is directly fixed by this, but not having pi.size correct is
clearly a pain for debugging, and it's definitely calling for trouble in
the long run.