add some debug prints when rolling back sql transactions
We have multiple thousand transactions in the python stress-test
testcase, and about a hundred lead to a rollback (and a few lead to
more). This doesn't seem outrageously bad.

It's true however that this testcase takes a much longer time on a big
iron than on my laptop, and I don't understand why.