Fix command line in gdb tests.
Apparently there's an inconsistency in the gdb command line: specifying
the auto-load safe path _on the command line_ doesn't seem to work
because apparently the file gets read early on. It works better to
postpone the binary file read to within the set of commands that are
processed from the command line.