Rework logapprox / lognorm_smart
As regards #30107, we have a bit of a dilemma.

Doing ulp-level computations to determine if we have the expected sign
changes is in fact extremely fragile, so we can't do it. False positives
would make the logapprox step too expensive, too often.

Going to long doubles when we think we need it isn't really providing a
satisfactory fix for the problem raised by #30107. It does fix the
inaccuracy bug, but just by sweeping it under the rug. It does _not_ fix
the problem with all compilers/architectures, though.

The solution that just validates an interval once it has been created,
by doing the sanity checks that all things compare as they should, looks
a bit better.