replace buggy fb_root_in_qlattice_127bits by a generic alternative
The whole point of this function is to be used when the special-q (and
hence the q-lattice basis) exceeds the usual bounds, which happens in
the descent. E.g. when the basis coefficients exceed 31 bits. It's a bit
of a pity, then, that it's implicitly making the assumption that the
basis coefficients are below 31 bits....

In order to "properly" fix this, we can either:
 - reimport existing code for redc_64 and invmod_redc_64, possibly fix
   it, expand the type of redc_invp_t to 64 bits, and expose all of that
   if SUPPORT_LARGE_Q is defined
 - or use generic code based on mpzs. Since it's in the descent only,
   it's expensive, yes, but not necessarily catastrophic.
and for the time being, let's choose the latter.