Fix leak: kill wrapped_modint_t
This type is an example of the awful traps of macro-based polymorphism.
If I create a single wrapped_modint_t, with modint_t in itself that
resolves to multiple types in different compilation units, then linking
is allowed to do whatever it wants, and keep only one instantiation. For
better or for worse.

We can't template on the actual modint_t type either, since we don't
have type overloads for it either.