simplify primality testing in arithxx
It turns out that a Miller-Rabin test in base 2 + a Lucas V strong test
(with Q=1 and P smallest odd) has no counterexample below 2^64. The
latter is about twice as expensive as the MR test (one square + one mul
per bit, compared to one square and one addition), but it is anyway less
than the multiple bases we used to do, and has negligible impact anyway
since it is *behind* the MR test.

Overall this should have zero impact on speed, but it does simplify
code.