Parallel descent.
This is actually very tricky. The descent tree is a mutating tree, as we
constantly add nodes, replay some, and possibly abandone some with a
hope of rolling back.

This commit introduces a data structure that makes it possible to do
this in parallel, i.e. with multiple threads able to handle multiple
special-q's that appear in the same descent tree.

It's still WIP.

Currently (as a test) we're failing quite eagerly. The old mechanism of
increasing A and mfb until we get something can be put back in place
easily.

The present code is also printing-happy.