Templatize matpoly, + accept the idea that we'll require c++20.
The original driver for this heavy change is the fact that newer lcov
screams at our (admittedly ugly) conditional compilation of functions
based on preprocessor macros. While some of these are marginal
modifications, the one that reacts on `LINGEN_BINARY` is actually a very
pervasive one. This commit gets rid of all of it, and creates several
pairs of distinct types, based on the boolean is_binary. It all begins
with `matpoly<is_binary>`, for instance.

Most of this can be done with reasonably standard template stuff, with
nothing extremely fancy. There are cases where an `if constexpr` saves a
great deal of red tape though, and given that I've been holding off the
c++20 transition for some time now, I think that now is the time.
(Technically, `if constexpr` is c++17 though -- as it turns out there
are quite a few post-c++11 features that we've already been using for
some time, such as enable_if_t). There's no c++20 change in this commit
beyond these `if constexpr`.

This commit _almost_ gets rid of the ambiguously named identifiers. In
toplevel binaries, we do have preprocessor-triggered specific
instantiations, but that's sort of unavoidable (if we want distinct
binaries). One remaining offender is the `arith_hard` typedef, whose
target type still depends on the preprocessor macros.