Some binaries in CMakeLists-nodist.txt don't compile...
In a way, it's fair game -- this code is here for example purposes
only, and we do not use it in any way.

A side effect is that we don't care if these binaries _work_, at the
moment. IMHO, we whould rather have this code AND a test, or not have it
at all if we have no test for it.