bump timeout for test_generate_strategies
This test seems to go beyond the timeout every now and
then. This whole stack of code is still in need of a total rewrite, so
let's not worry too much about what's going on currently. The timeout
was at its default value of 20 seconds, for a typical run on an unloaded
machines that is in the whereabouts of 3 seconds. I'm not that much
surprised that on a loaded runner, it can drift significantly.